requirements.txt
GitPython==3.1.31
nltk==3.4.5
scikit_learn==1.2.2
networkx~=2.4
scikit-learn~=0.22.1